fan components with a deck spread without styles
