+++ src/app/components/TarotCard/TarotCardStyles.css
@@ -1,18 +1,27 @@
 /*TarotCardStyles.css*/
+
+
 .card {
     position: absolute;
-    width: 100px;
-    height: 150px;
+    width: 140px;
+    height: 140px;
     cursor: pointer;
     transition: transform 0.5s;
+    margin: 30rem 35rem;
+
+
+    transform-style: preserve-3d;
+    transition: transform 0.5s;
+
   }
   
-  .card .card-inner {
-    width: 100%;
-    height: 100%;
+  .card-inner {
+
+  
     transform-style: preserve-3d;
     transition: transform 0.5s;
   }
+
   
   .card.flipped .card-inner {
     transform: rotateY(180deg);
@@ -42,8 +51,8 @@
     justify-content: center;
   } */
   
-  /* .tirada {
+  .tirada {
     display: flex;
     justify-content: center;
   }
-   */
+
