+++ src/app/components/TarotCardFan/TarotCardFan.js
@@ -1,9 +1,11 @@
- //TarotCardFan.js
+
+ // TarotCardFan se encargará de la representación visual de una carta en una posición específica.
+
  import React from 'react';
 
  const TarotCardFan = ({ x, y, angle, card, onCardClick }) => (
    <div
-     className={`card absolute z-10 w-16 h-24 bg-blue-500 rounded-md border border-black`}
+     className={`card absolute z-150 w-100 h-24 bg-blue-500 rounded-md border border-black`}
      style={{
        left: `${y}px`,
        bottom: `${x}px`,
