+++ src/app/globals.css
@@ -25,3 +25,4 @@ body {
     )
     rgb(var(--background-start-rgb));
 }
+
