+++ src/app/page.js
@@ -3,16 +3,14 @@ import React, { useState, useEffect } from 'react';
 import TarotCard from './components/TarotCard/TarotCard';
 import TarotMeaning from './components/TarotMeaning/TarotMeaning';
 import tarotCardsData from '../app/data/sakuraCards.json'
-import TarotCardFan from './components/TarotCardFan/TarotCardFan';
-import styles from '../app/globals.css';
+import TarotCardCircle from './components/TarotCardCircle/TarotCardCircle';
+import styles from './components/TarotCard/TarotCardStyles.css';
 
 const Page = () => {
   const [selectedCards, setSelectedCards] = useState([]);
   const [meanings, setMeanings] = useState([]);
-  const [cards, setCards] = useState([]);
 
   const NUM_CARDS = 55;
-  const CARD_SPACING = 0.09;
   const ARC_RADIUS = 50;
 
   const handleCardClick = (card) => {
@@ -32,48 +30,29 @@ const Page = () => {
     return shuffledArray;
   };
 
-  useEffect(() => {
-    const calculateCoords = (numCards, arcRadius, cardSpacing) => {
-      const coords = [];
-
-      for (let i = 0; i < numCards; i++) {
-        const angle = (i / (numCards - 1)) * Math.PI - Math.PI / 2;
-        const x = arcRadius * Math.cos(angle);
-        const y = arcRadius * Math.sin(angle);
-        coords.push({ x, y, angle: (angle * 180) / Math.PI });
-      }
-
-      return coords;
-    };
-
-    const coords = calculateCoords(NUM_CARDS, ARC_RADIUS, CARD_SPACING);
-
-    setCards(
-      coords.map((coord, index) => (
-        <TarotCardFan
-          key={index}
-          card={tarotCardsData[index]}
-          onCardClick={handleCardClick}
-          x={coord.x}
-          y={coord.y}
-          angle={coord.angle}
-        />
-      ))
-    );
-  }, [handleCardClick]);
-
   return (
     <div className="container-cards">
-      <div className={`grid ${styles.grid}`}>
-        {cards}
+      <div className='container-tarot-circle'>
+        <TarotCardCircle
+          numCards={NUM_CARDS}
+          arcRadius={ARC_RADIUS}
+          cardSpacing={0.001}
+          tarotCardsData={tarotCardsData}
+          onCardClick={handleCardClick}
+        />
       </div>
       <div className={`tirada ${styles.tirada}`}>
-        {selectedCards.map((card, index) => (
-          <div key={card.id} className="flex justify-center items-center space-x-4">
-            <TarotCard card={card} />
-            <TarotMeaning meaning={meanings[index]} />
-          </div>
-        ))}
+        <section className="tirada-container flex">
+          {selectedCards.map((card, index) => (
+            <div key={card.id} className='relative'>
+              <TarotCard card={card} className='flex '/> {/* Ajusta la clase según tu diseño */}
+              <TarotMeaning meaning={meanings[index]} />
+            </div>
+          ))}
+        </section>
+        <section>
+          {/* Aquí puedes agregar cualquier otro contenido que desees mostrar */}
+        </section>
       </div>
     </div>
   );
